+++ clang-tools-extra/test/clang-tidy/infrastructure/Inputs/experimental-header-filter-matching/macro_def_header.h
@@ -1,0 +1,6 @@
+#ifndef MACRO_DEF_HEADER_H
+#define MACRO_DEF_HEADER_H
+
+#define DEFINE_CONFUSABLE(Name) int Name = 0;
+
+#endif