+++ clang-tools-extra/test/clang-tidy/infrastructure/Inputs/experimental-header-filter-matching/macro_expansion_header.h
@@ -1,0 +1,3 @@
+#include "macro_def_header.h"
+
+DEFINE_CONFUSABLE(l0)