+++ clang-tools-extra/test/clang-tidy/infrastructure/Inputs/experimental-header-filter-matching/wrapper_header.h
@@ -1,0 +1,3 @@
+namespace ns {
+#include "wrapper_kept_header.h"
+}